Revert "Improving router logic, some minor housekeeping to prepare for future features"
