+++ package.json
@@ -11,12 +11,9 @@
   "scripts": {
     "react-dev": "webpack --watch",
     "server-dev": "nodemon server/index.js",
-    "build": "webpack",
-    "test": "jest",
-    "debug-test": "jest -i --detectOpenHandles"
+    "build": "webpack"
   },
   "dependencies": {
-    "dotenv": "^8.2.0",
     "enzyme": "^3.11.0",
     "express": "^4.17.1",
     "jest": "^26.6.3",
@@ -35,10 +32,8 @@
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jquery": "^3.5.1",
     "nodemon": "^2.0.7",
-    "puppeteer": "^5.5.0",
     "react-dom": "^17.0.1",
     "style-loader": "^2.0.0",
-    "supertest": "^6.1.1",
     "terser-webpack-plugin": "^5.1.1",
     "webpack": "^5.13.0",
     "webpack-cli": "^4.3.1"
