--- sample.env
@@ -1,2 +1,0 @@
-DEV_PORT=4000
-DEV_DB=42127
