+++ server/db/index.js
@@ -1,6 +1,5 @@
 const mongoose = require ('mongoose');
-require('dotenv').config()
-mongoose.connect(`mongodb://localhost/${process.env.DEV_DB}`, { useNewUrlParser: true, useUnifiedTopology: true });
+mongoose.connect('mongodb://localhost/21027', { useNewUrlParser: true, useUnifiedTopology: true });
 
 const { seed } = require('./data_seed.js');
 
@@ -20,10 +19,6 @@ let retrieve = (filter) => {
   return Repo.find(filter, (err, docs) => err ? err : docs);
 }
 
-let remove = (filter) => {
-  return Repo.deleteMany(filter, (err) => err);
-}
-
 let seeder = () => {
   let w_seed = seed();
   for (let item of w_seed) {
@@ -37,10 +32,5 @@ let seeder = () => {
   }
 }
 
-let initializer = () => {
-  remove({})
-    .then(() => seeder());
-}
-
-module.exports.initializer = initializer;
+module.exports.seeder = seeder;
 module.exports.retrieve = retrieve;
